Rename package experimental to x (#462)
