+++ span_context.go
@@ -196,7 +196,7 @@ func (c SpanContext) IsFirehose() bool {
 }
 
 // ExtendedSamplingState returns the custom state object for a given key. If the value for this key does not exist,
-// it is initialized via initValue function. This state can be used by samplers (e.g. experimental.PrioritySampler).
+// it is initialized via initValue function. This state can be used by samplers (e.g. x.PrioritySampler).
 func (c SpanContext) ExtendedSamplingState(key interface{}, initValue func() interface{}) interface{} {
 	return c.samplingState.extendedStateForKey(key, initValue)
 }
